all srun overlap patch
(Craig: As I discussed with Cameron on July 19,
this is left separate to ease merging.)

this applies the "allow srun overlap" patch
that adds the "-s" to the right places, as
of the late June 2021 change.  We thought this
was in the repo as of the June 28 changes, but
this part appears to not be.  Without this
minor change, the servers don't start on
catalyst, but with it, they do.